Went off the reservation
There's a lot of work here but hopefully it is in the right
direction.
Separated Hub and Lightbulb into their own projects, they're the
demo framework.
Moved Devices library to Components libraries and added a section
for Communications between the Hub and Devices.
There is still a lot of work to be done on the Comms library.